Chapter 3 of Beck. At this point we identify that the Dollar is a value object - i.e. it uses that form of pattern. As a result a Value Object never changes once it has been constructed. However - it implies that it has to have an "equals" operator. The implementation of the test for equality throws up some thoughts that allows the concept of Triangulation to be introduced.
You can use triangulation to help form the tests/thoughts that are needed before you generalise the code. It helps you focus on the problem when the solution may not be clear/obvious.
The question asked is: What axes of variability are you trying to support in your design?
It is a way of approaching the problem from a different perspective.

Additionally added the "TODO" within the code - though will re-order it later.