CH4 of Beck (Privacy) In this case - we adjust a test to be more desriptive (after a fashion). We want it to say that the "times" operation returns an object, rather than just an integer - so the assertion can be against a Dollar Object, not just the int.
Once we have improved the test - we notice that the equality and multiplication tests are important dependent on each other. This is always a rish - but we have code that we wrote - and the test suite is separate - so it manages the risk - after a fashion.
We carry on!

We make a variable private - so that we can decouple it from the tests - i.e. we don't have to have access to the "amount" variable from within the test so that we make it private. This decoupling is an important principle to practice.