CH5 of Beck - Franc-ly speaking
Here we look at the first test on the proposed list - but it seems too large to write in one small go - so we break it down and look at multiplication of CHF first - and omit the details about rates.

Then we have a
a) test - the test FrancMultuplication
b) We made it compile
c) We run it to see that it fails
d) We make it run
e) Remove duplicaiton.

In this case (i.e. at this point in time) - we have not removed duplication - so we add it to our list, and carry on.

The key is that the small test that we created - represented progress.