CH 12 - Addition, Finally
In this chapter, we take a big test - which is addition cross-currencies, and reduce it to a simpler case of addition in a single currency.

Wiht that - we have a metaphor - Expressions - which are currency independent - think of them like a wallet.  Operations result in expressions - and if we think of the multiple ways that a currency combination in a wallet can have the same value - we introduce a "bank" as a way of converting to a single currency value at the end.

This means - Expression will operation on Money.
Then create a bank.
Have the bank reduce the value to USD (for example)